changed timeline to float

+++ Sources/NewmiSharedDTO/PnChannelsRegisterRq.swift
@@ -12,9 +12,9 @@ public struct PnChannelsRegisterRq: Codable, Sendable{
     public var pnToken: String?
     public var channels: [String]?
     public var deviceOS: String?
-    public var timeLine: [Int]?
+    public var timeLine: [CGFloat]?
     
-    public init(deviceIdentif: UUID? = nil, pnToken: String? = nil, channels: [String]? = nil, deviceOS: String? = nil, timeLine: [Int]? = nil) {
+    public init(deviceIdentif: UUID? = nil, pnToken: String? = nil, channels: [String]? = nil, deviceOS: String? = nil, timeLine: [CGFloat]? = nil) {
         self.deviceIdentif = deviceIdentif
         self.pnToken = pnToken
         self.channels = channels
